added missing tests
