+++ src/components/shop/shop-confirm-screen.js
@@ -8,7 +8,7 @@ import RexUIPlugin from "../../../lib/rexuiplugin.min.js";
 import { onScaleChange } from "../../core/scaler.js";
 import { Screen } from "../../core/screen.js";
 import { setBalance } from "./balance.js";
-import { createConfirm } from "./confirm.js";
+import { createConfirm } from "./confirm/confirm.js";
 import { initResizers } from "./backgrounds.js";
 
 export class ShopConfirm extends Screen {
