+++ test/components/shop/confirm.test.js
@@ -5,7 +5,7 @@
  * @license Apache-2.0
  */
 
-import { createConfirm } from "../../../src/components/shop/confirm.js";
+import { createConfirm } from "../../../src/components/shop/confirm/confirm.js";
 import * as layout from "../../../src/components/shop/shop-layout.js";
 import * as text from "../../../src/core/layout/text-elem.js";
 import * as buttons from "../../../src/components/shop/menu-buttons.js";
