+++ themes/default/shop-confirm/config.json5
@@ -40,7 +40,7 @@
             key: "shop-menu.menuButtonBackground",
             styles: { fontStyle: "bold", fontSize: "16px" },
         },
-        background: { key: "halfViewBackground", columns: [100, null, 100], rows: [100, null, 100] },
+        background: { key: "halfViewBackground", columns: [100, null, 100], rows: [100, null, 25] },
         detailView: true,
         prompt: {
             buy: {
